3d23615c8059baadfed043aeda3f67a5ec67a64d fix: handle multi byte keytypes in Key::decode and Key::encode (nymius)
d47ab6d18f2478fc79c5f4601279cf1e4657c009 fix: encode keytype as u64 instead of u8 (nymius)

Pull request description:

  `<keytype>` as specified by BIP 174 may be represented with up to 8 bytes, even if minimally encoded.

  Because the currently specified fields weren't using as many bytes, we were only considering the single byte case in the `Key::decode` and `Key::encode` logic.

  Here we change this logic to address larger `<keytypes>` and follow the specification.

  Stacks on top of: #90, #93.
  Fixes #78.

ACKs for top commit:
  nyonson:
    ACK 3d23615c8059baadfed043aeda3f67a5ec67a64d
  tcharding:
    ACK 3d23615c8059baadfed043aeda3f67a5ec67a64d

Tree-SHA512: f6d12532eabd702979020835f23f7e588620014f8850e15d02dcbea20f253eea894f5a9f36b95d66a55f0e38a3b6a6701e8ca06aff88f6c70eb57600d59c9ef1

+++ src/v0/bitcoin/map/global.rs
@@ -13,28 +13,27 @@ use crate::v0::bitcoin::map::Map;
 use crate::v0::bitcoin::{raw, Error, Psbt};
 
 /// Type: Unsigned Transaction PSBT_GLOBAL_UNSIGNED_TX = 0x00
-const PSBT_GLOBAL_UNSIGNED_TX: u8 = 0x00;
+const PSBT_GLOBAL_UNSIGNED_TX: u64 = 0x00;
 /// Type: Extended Public Key PSBT_GLOBAL_XPUB = 0x01
-const PSBT_GLOBAL_XPUB: u8 = 0x01;
+const PSBT_GLOBAL_XPUB: u64 = 0x01;
 /// Type: Transaction Version PSBT_GLOBAL_TX_VERSION = 0x02
-const PSBT_GLOBAL_TX_VERSION: u8 = 0x02;
+const PSBT_GLOBAL_TX_VERSION: u64 = 0x02;
 /// Type: Fallback Locktime PSBT_GLOBAL_FALLBACK_LOCKTIME = 0x03
-const PSBT_GLOBAL_FALLBACK_LOCKTIME: u8 = 0x03;
+const PSBT_GLOBAL_FALLBACK_LOCKTIME: u64 = 0x03;
 /// Type: Input Count PSBT_GLOBAL_INPUT_COUNT = 0x04
-const PSBT_GLOBAL_INPUT_COUNT: u8 = 0x04;
+const PSBT_GLOBAL_INPUT_COUNT: u64 = 0x04;
 /// Type: Output Count PSBT_GLOBAL_OUTPUT_COUNT = 0x05
-const PSBT_GLOBAL_OUTPUT_COUNT: u8 = 0x05;
+const PSBT_GLOBAL_OUTPUT_COUNT: u64 = 0x05;
 /// Type: Transaction Modifiable Flags PSBT_GLOBAL_TX_MODIFIABLE = 0x06
-const PSBT_GLOBAL_TX_MODIFIABLE: u8 = 0x06;
+const PSBT_GLOBAL_TX_MODIFIABLE: u64 = 0x06;
 /// Type: Version Number PSBT_GLOBAL_SP_ECDH_SHARE = 0x07
-const PSBT_GLOBAL_SP_ECDH_SHARE: u8 = 0x07;
+const PSBT_GLOBAL_SP_ECDH_SHARE: u64 = 0x07;
 /// Type: Version Number PSBT_GLOBAL_SP_DLEQ = 0x08
-const PSBT_GLOBAL_SP_DLEQ: u8 = 0x08;
+const PSBT_GLOBAL_SP_DLEQ: u64 = 0x08;
 /// Type: Version Number PSBT_GLOBAL_VERSION = 0xFB
-const PSBT_GLOBAL_VERSION: u8 = 0xFB;
-
+const PSBT_GLOBAL_VERSION: u64 = 0xFB;
 /// Type: Proprietary Use Type PSBT_GLOBAL_PROPRIETARY = 0xFC
-const PSBT_GLOBAL_PROPRIETARY: u8 = 0xFC;
+const PSBT_GLOBAL_PROPRIETARY: u64 = 0xFC;
 
 impl Map for Psbt {
     fn get_pairs(&self) -> Vec<raw::Pair> {
